feat: enable independent GQI recalculation
