+++ meg_qc/calculation/meg_qc_pipeline.py
@@ -43,447 +43,8 @@ import os
 import json
 import pandas as pd
 from typing import Union, Optional, Dict
-import plotly.graph_objects as go
-from plotly.offline import plot
-from statistics import mean
 
-def create_summary_report(
-    json_file: Union[str, os.PathLike],
-    html_output: str = None,
-    json_output: str = "first_sight_report.json",
-    gqi_settings: Optional[Dict[str, Dict[str, float]]] = None,
-):
-    """Create a human readable QC summary from the metrics JSON file.
-
-    Parameters
-    ----------
-    json_file : str or Path
-        Path to the intermediate metrics JSON generated by the pipeline.
-    html_output : str | None
-        Optional HTML output path. If ``None`` only ``json_output`` is written.
-    json_output : str
-        Path of the JSON summary that will be created.
-    gqi_settings : dict | None
-        Dictionary with thresholds and weights for the Global Quality Index.
-        It may also contain ``compute_gqi`` and ``include_ecg_eog`` booleans
-        to control the index computation.
-    """
-    # === Load JSON ===
-    with open(json_file, "r", encoding="utf-8") as f:
-        data = json.load(f)
-
-    # Extract the total number of events evaluated in the muscle metric. When
-    # present, this allows expressing the number of detected artifacts as a
-    # percentage of all evaluated events rather than an absolute count.
-    total_events = data.get("MUSCLE", {}).get("total_number_of_events")
-
-    # PSD global: porcentaje de potencia en las frecuencias detectadas como ruido
-    psd_details_mag = (
-        data.get("PSD", {})
-        .get("PSD_global", {})
-        .get("mag", {})
-        .get("details", {})
-    )
-    psd_details_grad = (
-        data.get("PSD", {})
-        .get("PSD_global", {})
-        .get("grad", {})
-        .get("details", {})
-    )
-    noisy_power_mag = sum(
-        d.get("percent_of_this_noise_ampl_relative_to_all_signal_global", 0)
-        for d in psd_details_mag.values()
-    )
-    noisy_power_grad = sum(
-        d.get("percent_of_this_noise_ampl_relative_to_all_signal_global", 0)
-        for d in psd_details_grad.values()
-    )
-    M_psd = mean([noisy_power_mag, noisy_power_grad])
-
-    # Parameters for the GQI computation can be tuned in the ``thresholds``
-    # dictionary defined below. ``start`` and ``end`` define the linear range
-    # where the quality transitions from 100 % to 0 %, while ``weight`` controls
-    # the importance of each metric in the final index.
-
-    # Detect automatically if any metric used in the global report is missing
-    # or empty. This avoids failures when datasets are not suitable for all
-    # calculations.
-    metrics_to_check = ["STD", "PSD", "PTP_MANUAL", "MUSCLE"]
-    compute_gqi = True
-    include_corr = True
-    if gqi_settings is not None:
-        compute_gqi = gqi_settings.get("compute_gqi", True)
-        include_corr = gqi_settings.get("include_ecg_eog", True)
-    if compute_gqi and include_corr:
-        metrics_to_check.extend(["ECG", "EOG"])
-    missing = [m for m in metrics_to_check if not data.get(m)]
-    if missing:
-        print(
-            f"___MEGqc___: Skipping GlobalSummaryReport for {json_file}. "
-            f"Missing metrics: {', '.join(missing)}"
-        )
-        return
-    if html_output != None:
-        html_name = os.path.splitext(os.path.basename(json_output))[0].replace("-GlobalSummaryReport_meg", "")
-
-    # === STD TABLE (COMBINED VALUES) ===
-    def build_summary_table(source):
-        """Create a summary DataFrame listing noisy and flat channels."""
-        rows = []
-        for sensor_type in ["mag", "grad"]:
-            n_noisy = source[sensor_type]["number_of_noisy_ch"]
-            p_noisy = source[sensor_type]["percent_of_noisy_ch"]
-            n_flat = source[sensor_type]["number_of_flat_ch"]
-            p_flat = source[sensor_type]["percent_of_flat_ch"]
-            rows.append({
-                "Metric": "Noisy Channels",
-                sensor_type: f"{n_noisy} ({p_noisy:.1f}%)"
-            })
-            rows.append({
-                "Metric": "Flat Channels",
-                sensor_type: f"{n_flat} ({p_flat:.1f}%)"
-            })
-        df = pd.DataFrame(rows)
-        df = df.groupby("Metric").first().reset_index()
-        df.rename(columns={"mag": "MAGNETOMETERS", "grad": "GRADIOMETERS"}, inplace=True)
-        return df
-
-    general_df = build_summary_table(data["STD"]["STD_all_time_series"])
-    ptp_df = build_summary_table(data["PTP_MANUAL"]["ptp_manual_all"])
-
-    def build_psd_summary(noise_mag, noise_grad):
-        """Return a DataFrame summarising the percentage of noisy power."""
-        df = pd.DataFrame([
-            {
-                "Metric": "Noise Power",
-                "mag": noise_mag,
-                "grad": noise_grad,
-            }
-        ])
-        df["mag"] = df["mag"].map(lambda v: f"{v:.2f}%")
-        df["grad"] = df["grad"].map(lambda v: f"{v:.2f}%")
-        df.rename(columns={"mag": "MAGNETOMETERS", "grad": "GRADIOMETERS"}, inplace=True)
-        return df
-
-    psd_df = build_psd_summary(noisy_power_mag, noisy_power_grad)
-
-    # === GLOBAL QUALITY INDEX ===
-    # Ranges and weights used for GQI calculation. Start indicates the
-    # threshold below which quality is 100 %, end indicates 0 % quality,
-    # and weight controls the metric contribution to the final index.
-
-    # ``ch`` accounts for noisy/flat channels, ``corr`` for ECG/EOG
-    # contamination, and ``mus`` for muscle artifacts. Adjust the numbers if
-    # different data sets require stricter or looser rules.
-    if gqi_settings is not None:
-        thresholds = {
-            k: v for k, v in gqi_settings.items() if isinstance(v, dict)
-        }
-    else:
-        thresholds = {
-            # Weights sum to 1.0 including the PSD contribution.
-            "ch":   {"start": 5.0,  "end": 30.0, "weight": 0.32},
-            "corr": {"start": 5.0,  "end": 25.0, "weight": 0.24},
-            "mus":  {"start": 1.0,  "end": 10.0, "weight": 0.24},
-            "psd":  {"start": 1.0,  "end": 5.0,  "weight": 0.2},
-        }
-
-    # Utility converting a measured percentage ``M`` into a quality value
-    # between 0 and 1 using a progressive linear ramp. ``start`` marks the
-    # boundary for perfect quality, ``end`` for worst quality.
-    def quality_q(M, start, end):
-        """Map a metric value ``M`` onto a 0-1 quality scale."""
-        if M <= start:
-            return 1.0
-        elif M >= end:
-            return 0.0
-        else:
-            f = (M - start) / (end - start)
-            return 1.0 - f
-
-    # === CORRELATION ANALYSIS ===
-    # Build a small summary table for ECG/EOG contamination and return the
-    # percentage of channels exceeding the correlation threshold for each sensor
-    # type.
-    def count_high_correlations_from_details(section, contamination_key):
-        """Count channels with strong ECG/EOG correlation."""
-        results = []
-        percentages = []
-        for sensor_type in ["mag", "grad"]:
-            entries = (
-                data.get(section, {})
-                .get(contamination_key, {})
-                .get(sensor_type, {})
-                .get("details", {})
-            )
-            if not isinstance(entries, dict):
-                entries = {}
-            total = len(entries)
-            high_corr = sum(
-                1
-                for _, pair in entries.items()
-                if isinstance(pair, (list, tuple)) and pair and abs(pair[0]) > 0.8
-            )
-            percent = 100 * high_corr / total if total > 0 else 0
-            percentages.append(percent)
-            results.append(
-                {
-                    "Sensor Type": "MAGNETOMETERS" if sensor_type == "mag" else "GRADIOMETERS",
-                    "# |High Correlations| > 0.8": f"{high_corr} ({percent:.1f}%)",
-                    "Total Channels": total,
-                }
-            )
-        return pd.DataFrame(results), percentages
-
-    ecg_df, ecg_percents = count_high_correlations_from_details(
-        "ECG", "all_channels_ranked_by_ECG_contamination_level"
-    )
-    eog_df, eog_percents = count_high_correlations_from_details(
-        "EOG", "all_channels_ranked_by_EOG_contamination_level"
-    )
-
-    ecg_desc = str(data.get("ECG", {}).get("description", "")).lower()
-    eog_desc = str(data.get("EOG", {}).get("description", "")).lower()
-    def is_noisy(desc: str) -> bool:
-        noisy_markers = ["too noisy", "does not have expected", "can not be detected"]
-        return any(m in desc for m in noisy_markers)
-
-    ecg_noisy = is_noisy(ecg_desc)
-    eog_noisy = is_noisy(eog_desc)
-
-    # ---- Observed metrics ----
-    # Combine noisy and flat channel percentages across STD and PTP metrics to
-    # obtain a single estimate of how many sensors are unreliable.
-    bad_pct = mean([
-        data["STD"]["STD_all_time_series"]["mag"]["percent_of_noisy_ch"],
-        data["STD"]["STD_all_time_series"]["mag"]["percent_of_flat_ch"],
-        data["STD"]["STD_all_time_series"]["grad"]["percent_of_noisy_ch"],
-        data["STD"]["STD_all_time_series"]["grad"]["percent_of_flat_ch"],
-        data["PTP_MANUAL"]["ptp_manual_all"]["mag"]["percent_of_noisy_ch"],
-        data["PTP_MANUAL"]["ptp_manual_all"]["mag"]["percent_of_flat_ch"],
-        data["PTP_MANUAL"]["ptp_manual_all"]["grad"]["percent_of_noisy_ch"],
-        data["PTP_MANUAL"]["ptp_manual_all"]["grad"]["percent_of_flat_ch"],
-    ])
-
-    def mean_or_end(percs):
-        vals = [p for p in percs if p is not None]
-        return mean(vals) if vals else thresholds["corr"]["end"]
-
-    ecg_pct = mean_or_end(ecg_percents)
-    eog_pct = mean_or_end(eog_percents)
-
-    weight_corr_each = thresholds["corr"]["weight"] / 2
-    ecg_present = bool(data.get("ECG"))
-    eog_present = bool(data.get("EOG"))
-    weight_ecg = weight_corr_each if include_corr and ecg_present else 0.0
-    weight_eog = weight_corr_each if include_corr and eog_present else 0.0
-
-    if weight_ecg:
-        q_corr_ecg = 0.5 if ecg_noisy else quality_q(
-            ecg_pct, thresholds["corr"]["start"], thresholds["corr"]["end"]
-        )
-    else:
-        q_corr_ecg = 1.0
-
-    if weight_eog:
-        q_corr_eog = 0.5 if eog_noisy else quality_q(
-            eog_pct, thresholds["corr"]["start"], thresholds["corr"]["end"]
-        )
-    else:
-        q_corr_eog = 1.0
-
-    # Muscle contamination expressed as percentage of all evaluated events. If
-    # the total count is unavailable we fall back to the raw number of detected
-    # events.
-    muscle_events = data["MUSCLE"]["zscore_thresholds"]["number_muscle_events"]
-    if total_events:
-        muscle_pct = 100.0 * muscle_events / total_events
-    else:
-        muscle_pct = float(muscle_events)
-
-    # ---- Quality sub-indices ----
-    # Convert the observed percentages into values between 0 and 1 using the
-    # ramp function defined above.
-    q_ch = quality_q(bad_pct, thresholds["ch"]["start"], thresholds["ch"]["end"])
-    q_mus = quality_q(muscle_pct, thresholds["mus"]["start"], thresholds["mus"]["end"])
-    q_psd = quality_q(M_psd, thresholds["psd"]["start"], thresholds["psd"]["end"])
-
-    weights_sum = (
-        thresholds["ch"]["weight"]
-        + weight_ecg
-        + weight_eog
-        + thresholds["mus"]["weight"]
-        + thresholds["psd"]["weight"]
-    )
-    if weights_sum == 0:
-        weights_sum = 1.0
-
-    if compute_gqi:
-        GQI = round(
-            100.0
-            * (
-                thresholds["ch"]["weight"] * q_ch
-                + weight_ecg * q_corr_ecg
-                + weight_eog * q_corr_eog
-                + thresholds["mus"]["weight"] * q_mus
-                + thresholds["psd"]["weight"] * q_psd
-            )
-            / weights_sum,
-            2,
-        )
-        penalties = {
-            "ch": 100
-            * thresholds["ch"]["weight"]
-            * (1 - q_ch)
-            / weights_sum,
-            "corr": 100
-            * (
-                weight_ecg * (1 - q_corr_ecg)
-                + weight_eog * (1 - q_corr_eog)
-            )
-            / weights_sum,
-            "mus": 100
-            * thresholds["mus"]["weight"]
-            * (1 - q_mus)
-            / weights_sum,
-            "psd": 100
-            * thresholds["psd"]["weight"]
-            * (1 - q_psd)
-            / weights_sum,
-        }
-    else:
-        GQI = None
-        penalties = {}
-
-    # === EPOCH SUMMARY TABLE ===
-    def create_epoch_summary_table(source):
-        rows = []
-        for sensor_type, label in zip(["mag", "grad"], ["MAGNETOMETERS", "GRADIOMETERS"]):
-            n_noisy = source[sensor_type]["number_of_noisy_ch"]
-            p_noisy = source[sensor_type]["percent_of_noisy_ch"]
-            n_flat = source[sensor_type]["number_of_flat_ch"]
-            p_flat = source[sensor_type]["percent_of_flat_ch"]
-            rows.append({
-                "Sensor Type": label,
-                "Noisy Epochs": f"{n_noisy} ({p_noisy:.1f}%)",
-                "Flat Epochs": f"{n_flat} ({p_flat:.1f}%)"
-            })
-        return pd.DataFrame(rows)
-
-    std_epoch_df = create_epoch_summary_table(data["STD"]["STD_all_time_series"])
-    ptp_epoch_df = create_epoch_summary_table(data["PTP_MANUAL"]["ptp_manual_all"])
-
-    # === MUSCLE TABLE ===
-    muscle_events = data["MUSCLE"]["zscore_thresholds"]["number_muscle_events"]
-    muscle_df = pd.DataFrame([
-        {
-            "# Muscle Events": muscle_events,
-            "Total Events": total_events if total_events is not None else 0,
-        }
-    ])
-
-    # === HTML OUTPUT ===
-    std_lvl = data["STD"]["STD_all_time_series"]["mag"].get("std_lvl", "NA")
-    ptp_lvl = data["PTP_MANUAL"]["ptp_manual_all"]["mag"].get("ptp_lvl", "NA")
-    std_epoch_lvl = data["STD"]["STD_epoch"]["mag"].get("noisy_channel_multiplier", "NA")
-    ptp_epoch_lvl = data["PTP_MANUAL"]["ptp_manual_epoch"]["mag"].get("noisy_channel_multiplier", "NA")
-
-    style = """
-        <style>
-            body { font-family: Arial, sans-serif; margin: 10px; font-size: 16px; }
-            h1 { color: #003366; font-size: 25px; margin-bottom: 6px; font-weight: bold; }
-            h2 { color: #004d99; font-size: 19px; margin: 12px 0 6px 0; }
-            table { border-collapse: collapse; margin: 0 0 8px 0; font-size: 16px; }
-            th, td { border: 1px solid #ccc; padding: 6px 10px; text-align: center; }
-            th { background-color: #f2f2f2; }
-            .table-flex { display: flex; gap: 12px; flex-wrap: wrap; align-items: flex-start; margin-bottom: 12px; }
-            .table-box { flex: 1; min-width: 300px; }
-            .file-label { font-size: 18px; font-weight: bold; margin: 0 0 2px 12px; }
-            .subtitle { font-size: 19px; font-weight: bold; color: #222; margin: 0 0 12px 12px; }
-            .header-grid { display: grid; grid-template-columns: 1fr 1fr; align-items: start; margin-bottom: 0; }
-        </style>
-    """
-
-    if html_output != None:
-        with open(html_output, "w", encoding="utf-8") as f:
-            f.write("</div></div>")
-            f.write("<html><head><meta charset='UTF-8'>" + style + "</head><body>")
-            f.write("<div class='header-grid'>")
-            f.write("<div><h1>MEGQC Global Quality Report</h1></div>")
-            f.write(f"<div><div class='file-label'>File: {html_name}</div>")
-            f.write(f"<div class='subtitle'>Global Quality Index (GQI): {GQI}</div></div></div>")
-            f.write("<h2>GQI Penalties</h2>")
-            f.write(pd.DataFrame([
-                {"Metric": "Bad Channels", "Penalty (%)": f"{penalties['ch']:.2f}"},
-                {"Metric": "Correlation", "Penalty (%)": f"{penalties['corr']:.2f}"},
-                {"Metric": "Muscle", "Penalty (%)": f"{penalties['mus']:.2f}"},
-                {"Metric": "PSD Noise", "Penalty (%)": f"{penalties['psd']:.2f}"},
-            ]).to_html(index=False))
-            f.write("<div class='table-flex'>")
-            f.write(f"<div class='table-box'><h2>STD Time-Series (STD level: {std_lvl})</h2>")
-            f.write(general_df.to_html(index=False))
-            f.write(f"</div><div class='table-box'><h2>PTP Time-Series (STD level: {ptp_lvl})</h2>")
-            f.write(ptp_df.to_html(index=False))
-            f.write("</div></div>")
-            f.write("<h2>PSD Noise Summary</h2>")
-            f.write(psd_df.to_html(index=False))
-            f.write("<div class='table-flex'>")
-            f.write(f"<div class='table-box'><h2>STD Epoch Summary (STD level: {std_epoch_lvl})</h2>")
-            f.write(std_epoch_df.to_html(index=False))
-            f.write(f"</div><div class='table-box'><h2>PTP Epoch Summary (STD level: {ptp_epoch_lvl})</h2>")
-            f.write(ptp_epoch_df.to_html(index=False))
-            f.write("</div></div>")
-            f.write("<div class='table-flex'>")
-            f.write("<div class='table-box'><h2>ECG Correlation Summary</h2>")
-            f.write(ecg_df.to_html(index=False))
-            f.write("</div><div class='table-box'><h2>EOG Correlation Summary</h2>")
-            f.write(eog_df.to_html(index=False))
-            f.write("</div></div>")
-            f.write("<h2>Muscle Events Summary</h2>")
-            f.write(muscle_df.to_html(index=False))
-            f.write("</body></html>")
-
-    # === JSON SUMMARY OUTPUT ===
-    # json_output = os.path.splitext(html_output)[0] + "_summary.json"
-    file_name = os.path.basename(json_output)
-    summary_data = {
-        "file_name": file_name,
-        "GQI": GQI,
-        "STD_time_series": general_df.to_dict(orient="records"),
-        "PTP_time_series": ptp_df.to_dict(orient="records"),
-        "STD_epoch_summary": std_epoch_df.to_dict(orient="records"),
-        "PTP_epoch_summary": ptp_epoch_df.to_dict(orient="records"),
-        "ECG_correlation_summary": ecg_df.to_dict(orient="records"),
-        "EOG_correlation_summary": eog_df.to_dict(orient="records"),
-        "PSD_noise_summary": psd_df.to_dict(orient="records"),
-        "Muscle_events": {
-            "# Muscle Events": muscle_events,
-            "total_number_of_events": total_events,
-        },
-        "GQI_penalties": penalties,
-        "GQI_metrics": {
-            "bad_pct": bad_pct,
-            "ecg_pct": ecg_pct,
-            "eog_pct": eog_pct,
-            "muscle_pct": muscle_pct,
-            "psd_noise_pct": M_psd,
-        },
-        "parameters": {
-            "std_lvl": std_lvl,
-            "ptp_lvl": ptp_lvl,
-            "std_epoch_lvl": std_epoch_lvl,
-            "ptp_epoch_lvl": ptp_epoch_lvl
-        }
-    }
-
-    with open(json_output, "w", encoding="utf-8") as f_json:
-        json.dump(summary_data, f_json, indent=4)
-
-    print(f"HTML successfully generated: {html_output}")
-    print(f"JSON summary successfully generated: {json_output}")
-
-
-
+from meg_qc.calculation.metrics.summary_report_GQI import generate_gqi_summary
 
 def ctf_workaround(dataset, sid):
     artifacts = dataset.query(suffix="meg", return_type="object", subj=sid, scope='raw')
@@ -1325,14 +886,7 @@ def process_one_subject(
     # WRITE DERIVATIVE
     ancpbids.write_derivative(dataset, derivative)
 
-    # Generate Global report
-    json_simple_metrics = dataset.query(scope="derivatives/Meg_QC/calculation", descr="SimpleMetrics", suffix="meg",
-                                        extension="json")
-    for fname in json_simple_metrics:
-        json_path = fname.get_absolute_path()
-        if json_path:
-            output_json = json_path.replace('SimpleMetrics', 'GlobalSummaryReport')
-        create_summary_report(json_path, None, output_json, all_qc_params.get("GlobalQualityIndex"))
+
 
     # Removes intermediate trash objects
     del meg_artifact, derivative
@@ -1580,42 +1134,11 @@ def make_derivative_meg_qc(
                 for sub in excluded_subjects:
                     f.write(str(sub) + '\n')
 
-        # ---------------------------------------
-        # Create group level table with GQI values
-        # ---------------------------------------
+        # Generate Global Quality Index reports and group table
         try:
-            pattern = os.path.join(
-                dataset_path,
-                "derivatives",
-                "Meg_QC",
-                "calculation",
-                "sub-*",
-                "*GlobalSummaryReport_meg.json",
-            )
-            summary_files = glob.glob(pattern)
-            rows = []
-            for path in summary_files:
-                with open(path, "r", encoding="utf-8") as f:
-                    js = json.load(f)
-                subject = os.path.basename(os.path.dirname(path))
-
-                row = {"subject": subject}
-                row.update(flatten_summary_metrics(js))
-                rows.append(row)
-
-            if rows:
-                group_dir = os.path.join(
-                    dataset_path, "derivatives", "Meg_QC", "Group_metrics"
-                )
-                os.makedirs(group_dir, exist_ok=True)
-                df = pd.DataFrame(rows)
-                df.to_csv(
-                    os.path.join(group_dir, "Global_Quality_Index.tsv"),
-                    sep="\t",
-                    index=False,
-                )
+            generate_gqi_summary(dataset_path, config_file_path)
         except Exception as e:
-            print("___MEGqc___: Failed to create group metrics", e)
+            print("___MEGqc___: Failed to create global quality reports", e)
 
     return
