+++ setup.cfg
@@ -42,6 +42,7 @@ packages = find:
 include_package_data = True
 
 [options.entry_points]
-console_script=
+console_scripts=
     run-meg-qc = meg_qc.run_megqc:hello_world
+    globalqualityindex = meg_qc.test:run_gqi
